The spin projections table must contain integers only

+++ src/isingchat/ising.py
@@ -15,7 +15,7 @@ from .utils import bin_digits
 
 def make_spin_proj_table(num_neighbors: int):
     """Creates the table of spin projections."""
-    table = np.empty((2 ** num_neighbors, num_neighbors))
+    table = np.empty((2 ** num_neighbors, num_neighbors), dtype="i8")
     for idx in range(2 ** num_neighbors):
         projections = [-2 * v + 1 for v in bin_digits(idx, num_neighbors)]
         table[idx, :] = projections
